fix: frame address resolution on CPython 3.13
We fix a bug in the frame address resolution logic when reading a
frame from a local copy of a stack chunk.
